bugfix(zip): fix check for feature type presence
the geojson.line(features) call on a feature collection that does not
contain any line features returns an empty array nested inside an array
causing the l.geometries.length to always return true, which causes and
attempt to pefrorm a write on an empty array of geometries, which causes
the write call to never return, and thus keeps the zip function from finishing.

this is avoided by checking the length of the nested array instead.